Adding a mechanism to override tests in engine plugins from within the application, in a structured way.
In your engine tests, add the following line to the bottom of the
file:

  Engines::Testing.override_tests_from_app

This will check for files in the application's test directory, of the
name "test/{kind}/{plugin_name}/{test_name}, and load it to override
any tests. For example, if the test is in

  vendor/plugins/my_plugin/test/unit/stuff_test.rb

then the 'override_tests' method will load (if the file exists)

  test/unit/my_plugin/stuff_test.rb

Keeping these overrides namespaced by plugin should make it a bit
clearer what's going on.

Signed-off-by: Thincloud <admin@thincloud.com>